Foreigner now honors ignore in schemadummper
The SchemaDumper can be set to ignore tables -- so they are not written
out on a call to the SchemaDumper.

Prior to this change, Foreigner would write out *ALL* foreign keys,
even for tables that were not being processed.  With this change,
foreigner will now only process foreign keys on tables that are
being processed by the SchemaDumper.